backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-aiplatform
firebase-admin
elevenlabs
websockets